feat: moved all the pages to a designated folder
