+++ package.json
@@ -15,7 +15,9 @@
     "expo": "^46.0.21",
     "expo-status-bar": "~1.4.4",
     "react": "18.2.0",
-    "react-native": "0.71.8"
+    "react-native": "0.71.8",
+    "react-native-safe-area-context": "^4.5.3",
+    "react-native-screens": "^3.21.0"
   },
   "devDependencies": {
     "@babel/core": "^7.20.0"
